+++ CHANGELOG.md
@@ -5,7 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog], and this project adheres to [Semantic Versioning].
 
 ## Unreleased
-- /
+
+- added options to download sources and Javadoc via IDEA Gradle plugin
 
 ## [2.2.0] - 2026-06-18